Move `StacksString`
