+++ stackslib/src/util_lib/strings.rs
@@ -17,18 +17,14 @@
 use std::borrow::Borrow;
 use std::fmt;
 use std::io::{Read, Write};
-use std::ops::{Deref, DerefMut};
+use std::ops::Deref;
 
 use clarity::vm::errors::ClarityTypeError;
-use clarity::vm::representations::{
-    ClarityName, ContractName, MAX_STRING_LEN as CLARITY_MAX_STRING_LENGTH,
-};
+use clarity::vm::representations::MAX_STRING_LEN as CLARITY_MAX_STRING_LENGTH;
 use lazy_static::lazy_static;
 use regex::Regex;
-use stacks_common::codec::{
-    read_next, write_next, Error as codec_error, StacksMessageCodec, MAX_MESSAGE_LEN,
-};
-use stacks_common::util::retry::BoundReader;
+pub use stacks_codec::strings::StacksString;
+use stacks_common::codec::{read_next, write_next, Error as codec_error, StacksMessageCodec};
 use url;
 
 lazy_static! {
@@ -44,11 +40,6 @@ guarded_string!(
     ClarityTypeError::InvalidUrlString
 );
 
-/// printable-ASCII-only string, but encodable.
-/// Note that it cannot be longer than ARRAY_MAX_LEN (4.1 billion bytes)
-#[derive(Clone, PartialEq, Serialize, Deserialize)]
-pub struct StacksString(Vec<u8>);
-
 pub struct VecDisplay<'a, T: fmt::Display>(pub &'a [T]);
 
 impl<T: fmt::Display> fmt::Display for VecDisplay<'_, T> {
@@ -62,62 +53,6 @@ impl<T: fmt::Display> fmt::Display for VecDisplay<'_, T> {
             }
         }
         write!(f, "]")
-    }
-}
-
-impl fmt::Display for StacksString {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        f.write_str(String::from_utf8_lossy(self).into_owned().as_str())
-    }
-}
-
-impl fmt::Debug for StacksString {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str(String::from_utf8_lossy(self).into_owned().as_str())
-    }
-}
-
-impl Deref for StacksString {
-    type Target = Vec<u8>;
-    fn deref(&self) -> &Vec<u8> {
-        &self.0
-    }
-}
-
-impl DerefMut for StacksString {
-    fn deref_mut(&mut self) -> &mut Vec<u8> {
-        &mut self.0
-    }
-}
-
-impl StacksMessageCodec for StacksString {
-    fn consensus_serialize<W: Write>(&self, fd: &mut W) -> Result<(), codec_error> {
-        write_next(fd, &self.0)
-    }
-
-    fn consensus_deserialize<R: Read>(fd: &mut R) -> Result<StacksString, codec_error> {
-        let bytes: Vec<u8> = {
-            let mut bound_read = BoundReader::from_reader(fd, MAX_MESSAGE_LEN as u64);
-            read_next(&mut bound_read)
-        }?;
-
-        // must encode a valid string
-        let s = String::from_utf8(bytes.clone()).map_err(|_e| {
-            warn!("Invalid StacksString -- could not build from utf8");
-            codec_error::DeserializeError(
-                "Invalid Stacks string: could not build from utf8".to_string(),
-            )
-        })?;
-
-        if !StacksString::is_valid_string(&s) {
-            // non-printable ASCII or not ASCII
-            warn!("Invalid StacksString -- non-printable ASCII or non-ASCII");
-            return Err(codec_error::DeserializeError(
-                "Invalid Stacks string: non-printable or non-ASCII string".to_string(),
-            ));
-        }
-
-        Ok(StacksString(bytes))
     }
 }
 
@@ -169,65 +104,6 @@ impl StacksMessageCodec for UrlString {
             let _ = url.parse_to_block_url()?;
         }
         Ok(url)
-    }
-}
-
-impl From<ClarityName> for StacksString {
-    fn from(clarity_name: ClarityName) -> StacksString {
-        // .unwrap() is safe since StacksString is less strict
-        StacksString::from_str(&clarity_name).unwrap()
-    }
-}
-
-impl From<ContractName> for StacksString {
-    fn from(contract_name: ContractName) -> StacksString {
-        // .unwrap() is safe since StacksString is less strict
-        StacksString::from_str(&contract_name).unwrap()
-    }
-}
-
-impl StacksString {
-    /// Is the given string a valid Clarity string?
-    pub fn is_valid_string(s: &String) -> bool {
-        s.is_ascii() && StacksString::is_printable(s)
-    }
-
-    pub fn is_printable(s: &String) -> bool {
-        if !s.is_ascii() {
-            return false;
-        }
-        // all characters must be ASCII "printable" characters, excluding "delete".
-        // This is 0x20 through 0x7e, inclusive, as well as '\t' and '\n'
-        // TODO: DRY up with vm::representations
-        for c in s.as_bytes().iter() {
-            if (*c < 0x20 && *c != b'\t' && *c != b'\n') || *c > 0x7e {
-                return false;
-            }
-        }
-        true
-    }
-
-    pub fn is_clarity_variable(&self) -> bool {
-        ClarityName::try_from(self.to_string()).is_ok()
-    }
-
-    pub fn from_string(s: &String) -> Option<StacksString> {
-        if !StacksString::is_valid_string(s) {
-            return None;
-        }
-        Some(StacksString(s.as_bytes().to_vec()))
-    }
-
-    pub fn from_str(s: &str) -> Option<StacksString> {
-        if !StacksString::is_valid_string(&String::from(s)) {
-            return None;
-        }
-        Some(StacksString(s.as_bytes().to_vec()))
-    }
-
-    pub fn to_string(&self) -> String {
-        // guaranteed to always succeed because the string is ASCII
-        String::from_utf8(self.0.clone()).unwrap()
     }
 }
 
@@ -313,7 +189,7 @@ impl UrlString {
 
 #[cfg(test)]
 mod test {
-    use clarity::vm::representations::CONTRACT_MAX_NAME_LENGTH;
+    use clarity::vm::representations::{ClarityName, ContractName, CONTRACT_MAX_NAME_LENGTH};
 
     use super::*;
     use crate::net::codec::test::check_codec_and_corruption;
